moved header order into fields crate and in the process of moving the parser traits into the fields crate as well
