+++ src/wf_core/request.rs
@@ -1,6 +1,5 @@
-use super::wf_message_builder::FieldDefinitionParser;
 use std::ops::Mul;
-use wf_field::{definitions::*, Field, FieldDefinition};
+use wf_field::{definitions::*, Field, FieldDefinition, FieldDefinitionParserBase};
 
 const OBJECT_TYPE: FieldDefinition = Request::OBJECT_TYPE;
 const OBJECT_TYPE_QUANT: FieldDefinition = Request::OBJECT_TYPE_QUANT;
@@ -9,7 +8,7 @@ const OBJECT_TYPE_QUANT: FieldDefinition = Request::OBJECT_TYPE_QUANT;
 /// this function takes n number of request objects and parsers out the remaining request fields
 ///
 /// wf spec 4.3.1.9 Object Request Fields
-pub fn create_request_fields<T: FieldDefinitionParser>(parser: &mut T) -> Vec<Field> {
+pub fn create_request_fields<T: FieldDefinitionParserBase>(parser: &mut T) -> Vec<Field> {
     let ot_size = OBJECT_TYPE
         .expected_byte_length()
         .expect("request::OBJECT_TYPE is misconfigured: must have a start and end byte");
