+++ wf_field/src/header_order.rs
@@ -1,0 +1,28 @@
+use crate::{definitions::convert_value_to_code, Field};
+
+#[repr(usize)]
+#[derive(Copy, Clone)]
+pub enum MessageHeaderOrder {
+    Prefix = 0,
+    Version = 1,
+    EncryptionIndicator = 2,
+    DuressIndicator = 3,
+    MessageCode = 4,
+    ReferenceIndicator = 5,
+    ReferencedMessage = 6,
+}
+
+impl<'a> MessageHeaderOrder {
+    pub fn as_usize(&self) -> usize {
+        *self as usize
+    }
+
+    pub fn get<'b>(&'a self, fields: &'b [Field]) -> &'b Field {
+        &fields[self.as_usize()]
+    }
+
+    pub fn get_code(fields: &[Field]) -> (&Field, char) {
+        let field = Self::MessageCode.get(fields);
+        (field, convert_value_to_code(field.get()))
+    }
+}
