+++ wf_field/src/lib.rs
@@ -11,10 +11,14 @@ mod validation_test;
 pub mod definitions;
 mod field;
 mod field_definition;
+mod field_definition_parser;
+mod header_order;
 
 pub use definitions::{generic_header_fields, get_body_from_code, message_code, test_message_code};
 pub use field::Field;
 pub use field_definition::FieldDefinition;
+pub use field_definition_parser::{FieldDefinitionParser, FieldDefinitionParserBase};
+pub use header_order::MessageHeaderOrder;
 
 use definitions::get_body_from_code_char;
