+++ wf_parser/src/wf_header.rs
@@ -2,10 +2,10 @@ use wf_buffer::WhiteflagBuffer;
 use wf_field::{
     definitions,
     definitions::{convert_value_to_code, get_body_from_code},
-    Field, FieldDefinition,
+    Field, FieldDefinition, MessageHeaderOrder,
 };
 
-use super::{MessageHeader, MessageHeaderOrder};
+use super::MessageHeader;
 
 pub struct MessageHeaderValues {
     values: Vec<String>,
